tests/e2e/requirements.txt
pytest>=8.0
httpx>=0.27
psycopg2-binary>=2.9
paho-mqtt>=2.0
